updated to python-twitter 3.1 via git clone https://github.com/bear/python-twitter.git coordinates are working now

+++ Roadrunner/twitter_livestream/models.py
@@ -58,5 +58,5 @@ class Tweet(models.Model):
     user = models.ForeignKey(TUser, related_name='tweets')
     lang = models.CharField(max_length=10)
 
-    lat = models.IntegerField(null=True)
-    lon = models.IntegerField(null=True)
+    lat = models.FloatField(null=True)
+    lon = models.FloatField(null=True)
